refactor: 형광펜에 관한 api 파일명을 collection -> highlight로 변경
